python-metrics-app/requirements.txt
Flask==2.3.2
prometheus_flask_exporter==0.22.4
flask-opentracing==1.1.0
jaeger-client==4.8.0
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
